:bug: Fix a bug
user情報に関して、contextで参照していた値をrecoilの値に変更した
