--- src/PublicRoute.js
@@ -1,12 +1,0 @@
-import { Navigate } from 'react-router-dom';
-import { useAuthContext } from './AuthContext';
-
-const PublicRoute = ({ children }) => {
-    const { user } = useAuthContext();
-    if (user) {
-        return <Navigate to="/" />;
-    }
-    return children;
-};
-
-export default PublicRoute;
