+++ src/components/ui/deletButton/BlogDeleteButton.tsx
@@ -10,13 +10,13 @@ type Props = {
 const BlogDeleteButton: React.FC<Props> = ({ onClick }) => {
   let [isOpen, setIsOpen] = useState(false);
 
-  function closeModal() {
+  const closeModal = () => {
     setIsOpen(false);
-  }
+  };
 
-  function openModal() {
+  const openModal = () => {
     setIsOpen(true);
-  }
+  };
 
   return (
     <>
